+++ applications/launcher/controller.cpp
@@ -6,6 +6,7 @@
 #include <QDebug>
 #include <QDir>
 #include <QFile>
+#include <QSettings>
 #include <QStandardPaths>
 #include <fstream>
 #include <memory>
@@ -13,6 +14,349 @@
 
 QList<QString> configDirectoryPaths =
   {"/opt/etc/draft", "/etc/draft", "/home/root/.config/draft"};
+
+Controller::Controller(QObject* parent)
+  : QObject(parent)
+  , m_wifion(false)
+  , wifi("/sys/class/net/wlan0")
+  , applications() {
+  for (const auto& path : {
+         "/home/root/.config/oxide.conf",
+         "/opt/etc/oxide.conf",
+         "/etc/oxide.conf",
+       }) {
+    QFile file(path);
+    if (!file.exists()) {
+      continue;
+    }
+    qDebug() << "Migrating settings from" << path;
+    QSettings oxideConf(path, QSettings::IniFormat);
+    if (oxideConf.contains("columns")) {
+      sharedSettings.set_columns(oxideConf.value("columns").toInt());
+    }
+    if (oxideConf.contains("showWifiDb")) {
+      sharedSettings.set_showWifiDb(oxideConf.value("showWifiDb").toBool());
+    }
+    if (oxideConf.contains("showBatteryPercent")) {
+      sharedSettings.set_showBatteryPercent(
+        oxideConf.value("showBatteryPercent").toBool()
+      );
+    }
+    if (oxideConf.contains("showBatteryTemperature")) {
+      sharedSettings.set_showBatteryTemperature(
+        oxideConf.value("showBatteryTemperature").toBool()
+      );
+    }
+    if (oxideConf.contains("showDate")) {
+      sharedSettings.set_showDate(oxideConf.value("showDate").toBool());
+    }
+    if (oxideConf.contains("autoStartApplication")) {
+      sharedSettings.set_autoStartApplication(
+        oxideConf.value("autoStartApplication").toString()
+      );
+    }
+    file.remove();
+  }
+
+  networks = new WifiNetworkList();
+  notifications = new NotificationList();
+
+  auto bus = QDBusConnection::systemBus();
+  qDebug() << "Waiting for tarnish to start up";
+  while (!bus.interface()->registeredServiceNames().value().contains(
+    OXIDE_SERVICE
+  )) {
+    struct timespec args{
+      .tv_sec = 1,
+      .tv_nsec = 0,
+    },
+      res;
+    nanosleep(&args, &res);
+  }
+  qDebug() << "Requesting APIs";
+  api = new General(OXIDE_SERVICE, OXIDE_SERVICE_PATH, bus);
+  connect(api, &General::aboutToQuit, qApp, &QGuiApplication::quit);
+  auto reply = api->requestAPI("power");
+  reply.waitForFinished();
+  if (reply.isError()) {
+    qDebug() << reply.error();
+    qFatal("Could not request power API");
+  }
+  auto path = ((QDBusObjectPath)reply).path();
+  if (path == "/") {
+    qDebug() << "API not available";
+    qFatal("Power API was not available");
+  }
+  powerApi = new Power(OXIDE_SERVICE, path, bus);
+  // Connect to signals
+  connect(
+    powerApi,
+    &Power::batteryLevelChanged,
+    this,
+    &Controller::batteryLevelChanged
+  );
+  connect(
+    powerApi,
+    &Power::batteryStateChanged,
+    this,
+    &Controller::batteryStateChanged
+  );
+  connect(
+    powerApi,
+    &Power::batteryTemperatureChanged,
+    this,
+    &Controller::batteryTemperatureChanged
+  );
+  connect(
+    powerApi,
+    &Power::chargerStateChanged,
+    this,
+    &Controller::chargerStateChanged
+  );
+  connect(powerApi, &Power::stateChanged, this, &Controller::powerStateChanged);
+  connect(powerApi, &Power::batteryAlert, this, &Controller::batteryAlert);
+  connect(powerApi, &Power::batteryWarning, this, &Controller::batteryWarning);
+  connect(powerApi, &Power::chargerWarning, this, &Controller::chargerWarning);
+  reply = api->requestAPI("wifi");
+  reply.waitForFinished();
+  if (reply.isError()) {
+    qDebug() << reply.error();
+    qFatal("Could not request wifi API");
+  }
+  path = ((QDBusObjectPath)reply).path();
+  if (path == "/") {
+    qDebug() << "API not available";
+    qFatal("Wifi API was not available");
+  }
+  wifiApi = new Wifi(OXIDE_SERVICE, path, bus);
+  connect(wifiApi, &Wifi::disconnected, this, &Controller::disconnected);
+  connect(
+    wifiApi, &Wifi::networkConnected, this, &Controller::networkConnected
+  );
+  connect(wifiApi, &Wifi::stateChanged, this, &Controller::wifiStateChanged);
+  connect(wifiApi, &Wifi::rssiChanged, this, &Controller::wifiRssiChanged);
+  networks->setAPI(wifiApi);
+  auto state = wifiApi->state();
+  m_wifion = state != WifiState::WifiOff && state != WifiState::WifiUnknown;
+  QTimer::singleShot(1000, [this]() {
+    while (this->root == nullptr) {
+      qApp->processEvents(QEventLoop::AllEvents, 100);
+    }
+    // Get initial values when UI is ready
+    batteryLevelChanged(powerApi->batteryLevel());
+    batteryStateChanged(powerApi->batteryState());
+    batteryTemperatureChanged(powerApi->batteryTemperature());
+    chargerStateChanged(powerApi->chargerState());
+    powerStateChanged(powerApi->state());
+
+    wifiStateChanged(wifiApi->state());
+    wifiRssiChanged(wifiApi->rssi());
+    emit wifiOnChanged(m_wifion);
+    if (
+      stateController != nullptr && stateController->property("state") == "wifi"
+    ) {
+      connectWifiSignals();
+    }
+    auto network = wifiApi->network();
+    if (network.path() != "/") {
+      networkConnected(network);
+    }
+  });
+  reply = api->requestAPI("system");
+  reply.waitForFinished();
+  if (reply.isError()) {
+    qDebug() << reply.error();
+    qFatal("Could not request system API");
+  }
+  path = ((QDBusObjectPath)reply).path();
+  if (path == "/") {
+    qDebug() << "API not available";
+    qFatal("System API was not available");
+  }
+  systemApi = new System(OXIDE_SERVICE, path, bus);
+  connect(
+    systemApi,
+    &System::powerOffInhibitedChanged,
+    this,
+    &Controller::powerOffInhibitedChanged
+  );
+  connect(systemApi, &System::powerOffInhibitedChanged, [=](bool value) {
+    qDebug() << "Power Off Inhibited:" << value;
+  });
+  connect(
+    systemApi,
+    &System::sleepInhibitedChanged,
+    this,
+    &Controller::sleepInhibitedChanged
+  );
+  connect(systemApi, &System::autoSleepChanged, [this](int sleepAfter) {
+    setAutomaticSleep(sleepAfter);
+    setSleepAfter(sleepAfter);
+  });
+  connect(
+    systemApi, &System::swipeLengthChanged, [this](int direction, int length) {
+      setSwipeLength(direction, length);
+    }
+  );
+  auto autoSleep = systemApi->autoSleep();
+  setAutomaticSleep(autoSleep);
+  setSleepAfter(autoSleep);
+  auto autoLock = systemApi->autoLock();
+  setAutomaticLock(autoLock);
+  setLockAfter(autoLock);
+  for (short i = 1; i <= 4; ++i) {
+    setSwipeLength(i, systemApi->getSwipeLength(i));
+  }
+  emit powerOffInhibitedChanged(powerOffInhibited());
+  emit sleepInhibitedChanged(sleepInhibited());
+  reply = api->requestAPI("apps");
+  reply.waitForFinished();
+  if (reply.isError()) {
+    qDebug() << reply.error();
+    qFatal("Could not request apps API");
+  }
+  path = ((QDBusObjectPath)reply).path();
+  if (path == "/") {
+    qDebug() << "API not available";
+    qFatal("Apps API was not available");
+  }
+  appsApi = new Apps(OXIDE_SERVICE, path, bus);
+  connect(
+    appsApi,
+    &Apps::applicationUnregistered,
+    this,
+    &Controller::unregisterApplication
+  );
+  connect(
+    appsApi,
+    &Apps::applicationRegistered,
+    this,
+    &Controller::registerApplication
+  );
+  reply = api->requestAPI("notification");
+  reply.waitForFinished();
+  if (reply.isError()) {
+    qDebug() << reply.error();
+    qFatal("Could not request notification API");
+  }
+  path = ((QDBusObjectPath)reply).path();
+  if (path == "/") {
+    qDebug() << "API not available";
+    qFatal("Notification API was not available");
+  }
+  notificationApi = new Notifications(OXIDE_SERVICE, path, bus);
+  connect(
+    notificationApi,
+    &Notifications::notificationAdded,
+    this,
+    &Controller::notificationAdded
+  );
+  connect(
+    notificationApi,
+    &Notifications::notificationRemoved,
+    this,
+    &Controller::notificationRemoved
+  );
+  connect(
+    notificationApi,
+    &Notifications::notificationChanged,
+    this,
+    &Controller::notificationChanged
+  );
+  connect(
+    notifications,
+    &NotificationList::updated,
+    this,
+    &Controller::notificationsUpdated
+  );
+  connect(
+    notificationApi,
+    &Notifications::displayTimeChanged,
+    this,
+    [this](uint displayTime) { setNotificationDisplayTime(displayTime); }
+  );
+
+  reply = api->requestAPI("frontlight");
+  reply.waitForFinished();
+  if (!reply.isError()) {
+    path = ((QDBusObjectPath)reply).path();
+    if (path != "/") {
+      frontlightApi = new Frontlight(OXIDE_SERVICE, path, bus);
+      connect(
+        frontlightApi,
+        &Frontlight::extraBrightnessChanged,
+        this,
+        &Controller::extraBrightnessChanged
+      );
+    }
+  }
+
+  uiTimer = new QTimer(this);
+  uiTimer->setSingleShot(false);
+  uiTimer->setInterval(3 * 1000); // 3 seconds
+  connect(
+    uiTimer,
+    &QTimer::timeout,
+    this,
+    QOverload<>::of(&Controller::updateUIElements)
+  );
+  uiTimer->start();
+
+  // SharedSettings signal forwarding
+  connect(
+    &sharedSettings,
+    &Oxide::SharedSettings::columnsChanged,
+    this,
+    &Controller::columnsChanged
+  );
+  connect(
+    &sharedSettings,
+    &Oxide::SharedSettings::showWifiDbChanged,
+    this,
+    &Controller::showWifiDbChanged
+  );
+  connect(
+    &sharedSettings,
+    &Oxide::SharedSettings::showBatteryPercentChanged,
+    this,
+    &Controller::showBatteryPercentChanged
+  );
+  connect(
+    &sharedSettings,
+    &Oxide::SharedSettings::showBatteryTemperatureChanged,
+    this,
+    &Controller::showBatteryTemperatureChanged
+  );
+  connect(
+    &sharedSettings,
+    &Oxide::SharedSettings::showDateChanged,
+    this,
+    [this](bool showDate) {
+      if (root == nullptr) {
+        return;
+      }
+      QObject* clock = root->findChild<QObject*>("clock");
+      if (clock == nullptr) {
+        return;
+      }
+      QString text = "";
+      if (showDate) {
+        text = QDate::currentDate().toString(Qt::TextDate) + " ";
+      }
+      clock->setProperty(
+        "text", text + QTime::currentTime().toString("h:mm a")
+      );
+      emit showDateChanged(showDate);
+    }
+  );
+  connect(
+    &sharedSettings,
+    &Oxide::SharedSettings::autoStartApplicationChanged,
+    this,
+    &Controller::autoStartApplicationChanged
+  );
+}
+
 void
 Controller::loadSettings() {
   setNotificationDisplayTime(notificationApi->displayTime());
