+++ applications/launcher/launcher.pro
@@ -29,13 +29,8 @@ icons.files = ../../assets/opt/usr/share/icons/oxide/702x702/splash/oxide.png
 icons.path  = $$SPLASH_INSTALL_PATH
 INSTALLS += icons
 
-configFile.files = ../../assets/etc/oxide.conf
-configFile.path  = $$CONFIG_INSTALL_PATH
-INSTALLS += configFile
-
 DISTFILES += \
-    ../../assets/etc/dbus-1/system.d/org.freedesktop.login1.conf \
-    ../../assets/etc/oxide.conf
+    ../../assets/etc/dbus-1/system.d/org.freedesktop.login1.conf
 
 HEADERS += \
     controller.h \
